% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSeqlevels.R
\name{checkSeqlevels}
\alias{checkSeqlevels}
\title{Ensure that all seqlevel styles match across experiments}
\usage{
checkSeqlevels(x)
}
\arguments{
\item{x}{A \linkS4class{MultiAssayExperiment}}
}
\description{
Ensure that all seqlevel styles match across experiments
}
