% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSeqlevelsStyle.R
\name{checkSeqlevelsStyle}
\alias{checkSeqlevelsStyle}
\title{Ensure that all seqlevel styles match across experiments}
\usage{
checkSeqlevelsStyle(x)
}
\arguments{
\item{x}{A \linkS4class{MultiAssayExperiment}}
}
\description{
Ensure that all seqlevel styles match across experiments
}
