% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiOmicQC-checkers.R
\name{MultiOmicQC-checkers}
\alias{MultiOmicQC-checkers}
\alias{checkSeqlengths}
\alias{checkSeqlevels}
\alias{checkSeqlevelsStyle}
\title{Check seqinfo consistency across ranged experiments}
\usage{
checkSeqlengths(x)

checkSeqlevels(x)

checkSeqlevelsStyle(x)
}
\arguments{
\item{x}{A \linkS4class{MultiAssayExperiment}}
}
\description{
A set of functions to check several aspects of ranged
experiments including \code{seqlengths}, \code{seqlevels} and
\code{seqlevelsStyle}. These group of functions return a logical scalar
value when the names, styles, etc. are consistent accross all experiments.
}
\section{checkSeqlengths}{

Runs and checks \code{seqlengths} on all ranged experiments.
}

\section{checkSeqlevels}{

Checks \code{seqnames} levels of all ranged experiments for
consistency. Inconsistent seqlevels can be problematic when ranged
experiments differ by annotations.
}

\section{checkSeqlevelsStyle}{

Checks all \code{seqlevelsStyle} of all ranged experiments by comparing
the first style with the rest using the \code{duplicated} function.
}

