% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiOmicQC-pkg.R
\docType{package}
\name{MultiOmicQC-package}
\title{MultiOmicQC: Helper functions for checking the integrity of Multi-omics datasets}
\description{
MultiOmicQC allows the user to run common checks on MultiAssayExperiment objects
additional to the checks already established in MultiAssayExperiment.
}
\author{
\strong{Maintainer}: Marcel Ramos \email{marcel.ramos@roswellpark.org}

}
