% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiOmicQC-setters.R
\name{setSeqlevelsStyle}
\alias{setSeqlevelsStyle}
\title{Utility functions for setting metadata in MultiAssayExperiment}
\usage{
setSeqlevelsStyle(x, style)
}
\arguments{
\item{x}{A \linkS4class{MultiAssayExperiment}}

\item{style}{A character vector with a single element to specify the style}
}
\description{
Utility functions for setting metadata in MultiAssayExperiment
}
