% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSeqlevels.R
\name{checkSeqlevels}
\alias{checkSeqlevels}
\title{Ensure uniform seqnames across experiments}
\usage{
checkSeqlevels(x)
}
\arguments{
\item{x}{A \linkS4class{MultiAssayExperiment}}
}
\description{
This function checks seqnames of all ranged experiments for
consistency. This can become problematic when ranged experiments are
not using identical seqnames.
}
