% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiOmicQC-setters.R
\docType{methods}
\name{MultiOmicQC-setters}
\alias{MultiOmicQC-setters}
\alias{genome<-,MultiAssayExperiment-method}
\alias{`genome<-`}
\title{Utility functions for setting metadata in MultiAssayExperiment}
\usage{
\S4method{seqlevelsStyle}{MultiAssayExperiment}(x) <- value

\S4method{genome}{MultiAssayExperiment}(x) <- value
}
\arguments{
\item{x}{A \linkS4class{MultiAssayExperiment}}

\item{style}{A character vector with a single element to specify the style}
}
\description{
Utility functions for setting metadata in MultiAssayExperiment
}
